refactor: Folders structure
