+++ src/components/abstract/abstract-component.js
@@ -1,4 +1,4 @@
-import {createElement} from "./utils";
+import {createElement} from "../../utils";
 
 export default class AbstractComponent {
   constructor() {
