+++ src/components/films-list/page-layout.js
@@ -1,4 +1,4 @@
-import AbstractComponent from './abstract-component';
+import AbstractComponent from '../abstract/abstract-component';
 
 export default class PageLayout extends AbstractComponent {
   constructor() {
